refactor: ssa -> srg
