+++ src/analysis/ReachingDefinitions/Srg/CytronSRGBuilder.h
@@ -7,12 +7,12 @@
 #include <stack>
 
 #include "analysis/ReachingDefinitions/ReachingDefinitions.h"
-#include "analysis/ReachingDefinitions/Ssa/PhiPlacement.h"
+#include "analysis/ReachingDefinitions/Srg/PhiPlacement.h"
 
 namespace dg {
 namespace analysis {
 namespace rd {
-namespace ssa {
+namespace srg {
 
 /**
  * SparseRDGraphBuilder based on algorithm introduced in
